Update date / datetime validation due to github comments
