+++ docs/api-guide/fields.md
@@ -189,6 +189,8 @@ Corresponds to `django.db.models.fields.DateField`
 
 Uses `DATE_INPUT_FORMATS` to validate date.
 
+Optionally takes `format` as parameter to replace the matching pattern.
+
 ## DateTimeField
 
 A date and time representation.
@@ -196,6 +198,8 @@ A date and time representation.
 Corresponds to `django.db.models.fields.DateTimeField`
 
 Uses `DATETIME_INPUT_FORMATS` to validate date_time.
+
+Optionally takes `format` as parameter to replace the matching pattern.
 
 When using `ModelSerializer` or `HyperlinkedModelSerializer`, note that any model fields with `auto_now=True` or `auto_now_add=True` will use serializer fields that are `read_only=True` by default.
